requirements.txt
alembic==1.13.1
Flask==3.0.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
openai==0.27.0
psycopg2==2.8.6
psycopg2_binary==2.9.5
python_bcrypt==0.3.2
redis==5.0.4
SQLAlchemy==2.0.30
tenacity==8.2.2
tiktoken==0.7.0
zhipuai==2.1.0.20240521
